Fix linear form utilities, models and base
