+++ test/linear-form-utilities.jl
@@ -1,0 +1,14 @@
+using Test
+using SoleLogics
+using SoleModels
+
+
+b = Branch(LeftmostConjunctiveForm((@atoms p q r s)), "YES", "NO")
+
+@test_nowarn b[1:3]
+@test_nowarn b[[1]]
+@test_nowarn b[1]
+
+@test b[1:3] isa LeftmostConjunctiveForm
+@test b[[1]] isa LeftmostConjunctiveForm
+@test b[1] isa Atom
